moved from package screen to join_room_use_case
